WIP: isolated some AdhocMarkdown import logic into EntryImporter to enable live-reloading of markdown entries.
