+++ app/controllers/static_site/entries_controller.rb
@@ -38,6 +38,23 @@ module StaticSite
 
     private
     def set_entry
+
+      # TODO: live-reloading
+      # see if the file exists and if it does, import it
+
+      # step 1: does it exist as a file?
+      # step 2: is it markdown or yaml?
+      #   step 2.1: actually this is harder to untangle
+      #   will have to think about how i want to support the "normal" dump o yaml
+      #   vs the "adhoc" markdown
+      # step 3: parse it & add it.
+
+      @entry = EntryImporter.new(current_notebook).resolve_and_import!(params[:id])
+
+      if @entry
+        return
+      end
+
       # quick terrible hack for routing document type entries
       if params[:format]
         identifier = "#{params[:id]}.#{params[:format]}"
