+++ app/models/ad_hoc_markdown_importer.rb
@@ -54,20 +54,19 @@ class AdHocMarkdownImporter
 
       identifier = path_to_relative_identifier(file_path, notebook.import_path)
 
+      entry_importer = EntryImporter.new(current_notebook)
       if skip_file_path?(identifier, file_path)
         # Do nothing.
       elsif looks_like_text?(file_path)
         # Files that become normal entries, the content in our site.
         puts "processing #{file_path}" if Rails.env.development?
-        entry_attributes = entry_attributes_from_markdown(identifier, file_path)
-
-        entry = add_entry!(notebook, entry_attributes)
+        entry_importer.import!(identifier, file_path)
       else
         # everything that is not markdown is treated a bit differently.
         puts "processing #{file_path}" if Rails.env.development?
         entry_attributes = entry_attributes_from_document(identifier, file_path)
 
-        entry = add_entry!(notebook, entry_attributes)
+        entry = entry_importer.add_entry!(entry_attributes)
 
         filename = File.basename(identifier)
         if !entry.files.blobs.find_by(filename: filename)
@@ -92,71 +91,6 @@ class AdHocMarkdownImporter
 
   def looks_like_text?(file_path)
     file_path =~ /\.(md|markdown|html)$/
-  end
-
-  def entry_attributes_from_markdown(identifier, md_path)
-    loader = FrontMatterParser::Loader::Yaml.new(allowlist_classes: [Time, Date, DateTime])
-    md_parser = FrontMatterParser::SyntaxParser::Md.new
-    parsed_file = FrontMatterParser::Parser.new(md_parser, loader: loader).call(File.read(md_path))
-
-    occurred_at = parsed_file["occurred_at"]
-    if occurred_at.blank?
-      # let's try to guess it from the file
-      basename = File.basename(md_path)
-      date = basename.match(/([0-9]{4}-*[0-9]{2}-*[0-9]{2}-*)/).to_a[0]
-
-      # most of the time this does the right thing but it does have the habit
-      # of sometimes throwing an exception
-      begin
-        occurred_at = DateTime.parse(date.to_s)
-      rescue ArgumentError
-      end
-
-      # if still nil, let's look at the file itself
-      if occurred_at.nil?
-        occurred_at = File.ctime(md_path)
-      end
-    end
-
-    created_at = parsed_file["created_at"] || File.ctime(md_path)
-    updated_at = parsed_file["updated_at"] || File.mtime(md_path)
-
-    # if we're parsing front-mattered markdown, you don't get to define an
-    # identifier separate from the file's relative path, don't want to deal
-    # with collisions etc, too confusing, the ad hoc markdown is for ad hoc
-    # files, loosely slapped together!
-    #
-    # then, we lop off `.md`, `.markdown` and `html` from the suffix
-    entry_source = identifier # store the unmodified identifier as the "source"
-    identifier = identifier.gsub(/\.(md|markdown|html)/, "")
-
-    entry_attributes = parsed_file.front_matter.merge({
-      "identifier" => identifier,
-      "source" => entry_source,
-      "occurred_at" => occurred_at,
-      "body" => parsed_file.content,
-      "created_at" => created_at,
-      "updated_at" => updated_at,
-      skip_local_sync: true
-    }).slice(*Entry.accepted_attributes)
-
-    # handle metadata!
-    metadata_keys = (parsed_file.front_matter.keys - Entry.accepted_attributes)
-    if metadata_keys.any?
-      # if the user has specified a non Hash value, ie "metadata: foo", then
-      # throw a slightly easier to understand error here, instead of later on
-      # when we try to instantiate the Entry object & the error gets thrown there.
-      if entry_attributes["metadata"] && !entry_attributes["metadata"].is_a?(Hash)
-        raise "I expected the 'metadata' key on #{identifier} to be Hash, but something else is going on."
-      end
-
-      entry_attributes["metadata"] ||= {}
-      metadata_keys.each do |mkey|
-        entry_attributes["metadata"][mkey] ||= parsed_file.front_matter[mkey]
-      end
-    end
-
-    entry_attributes
   end
 
   def entry_attributes_from_document(identifier, file_path)
@@ -235,20 +169,5 @@ class AdHocMarkdownImporter
 
   def path_to_relative_identifier(file_path, import_path)
     Pathname.new(file_path).relative_path_from(import_path).to_s
-  end
-
-  def add_entry!(notebook, entry_attributes)
-    identifier = entry_attributes["identifier"]
-
-    # find or update the entry
-    entry = notebook.entries.find_by(identifier: identifier)
-
-    if entry
-      entry.update!(entry_attributes)
-    else
-      entry = notebook.entries.create(entry_attributes)
-    end
-
-    entry
   end
 end
